pyproject.toml
[tool.poetry]
name = "mailing-service"
version = "0.1.0"
description = "This is a service to create and manage mailing lists."
authors = ["Artur <RomanchenkoAS@yandex.ru>"]
readme = "README.md"
repository = "https://github.com/RomanchenkoAS/mailing_service"
keywords = ["mailing", "django-app"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.test.dependencies]
pytest = "*"
factory_boy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
